WEB-260 BrdImporter plugin
This plugin takes a testbench result and a .brd file that the user modified.
It runs BoardSynthesis.exe -r to import placement and routes from the .brd
into the layout.json file, then adds (or modifies) model elements in the
container so that the layout.json is used next Place and Route testbench run.